Fix change requests from Mike Murray
- One component per file
- Separate component and hoc container
- Remove link  from Footer component
