+++ imports/plugins/core/layout/client/components/footer.js
@@ -7,7 +7,6 @@ const Footer = () => (
   <div className="reaction-navigation-footer footer-default">
     <nav className="navbar-bottom" role="navigation">
       <div className="row">
-        <a href={Reaction.Router.pathFor("/")}>Home</a>
       </div>
     </nav>
   </div>
