+++ imports/plugins/core/ui-navbar/client/index.js
@@ -1,2 +1,3 @@
 export { default as Brand } from "./components/brand";
 export { default as Navbar } from "./containers/navbar";
+export { default as NavBarCheckout } from "./containers/navbarCheckout";
